speedscope format: make units conform to schema
Summary
-------

This commit replaces the string `"\u03bcs"` in the speedscope Units
enumeration with the string `"microseconds"`, so that the `"units"`
field of Speedscope-formatted Austin profiles conforms to the
Speedscope JSON schema defined in
https://github.com/jlfwong/speedscope/blob/main/src/lib/file-format-spec.ts.
Predecessor commits in this repo emit Speedscope-formatted JSON that
does not conform to the aforementioned JSON schema.

More details, background, and motivation
----------------------------------------

In predecessors to this commit, the `austin2speedscope` command line
tool would convert an Austin (CPU time) profile to speedscope format
such that the `profiles[i].units` field for any valid index of `i`
(e.g., 0) would contain the value `"\u03bcs"`, i.e., the UTF-8 code
point U+03BC -- displayed as the Greek letter "mu" in compatible
editors and terminals -- followed by the lowercase letter "s". While
this choice undoubtedly looks cool and displays as if one were
abbreviating the unit "microseconds" as in scientific and engineering
texts, this choice does not conform to the Speedscope JSON schema as
of upstream Speedscope commit
`e4040538376bbaa81f3746f11530717a1f2544b` (see
https://github.com/jlfwong/speedscope/commit/e4040538376bbaa81f3746f11530717a1f2544b1#diff-5bd093fe1cb8e10b3f1ae3621ec86c72191d17a90d73e4145518766b2d6d1ce3R117,
circa 2018-August), which has not since been changed as of this
writing. (For reference, as of this writing, the most recent upstream
commit is `e37f6fa7c38c110205e22081560b99cb89ce885e` and the relevant
link to the schema definition is at
https://github.com/jlfwong/speedscope/blob/e37f6fa7c38c110205e22081560b99cb89ce885e/src/lib/file-format-spec.ts#L117.)

Replacing the string `"\u03bcs"` with the string `"microseconds"`
does, however, conform to the Speedscope JSON schema, and results in
displaying the reformatted Austin profile in Speedscope's timeline
view with the proper units. Without this change, the Speedscope
timeline view omits units altogether. A viewer then has to simply know
that the horiztonal axis of the Speedscope icicle chart has units of
microsceconds, even though those units are omitted from the graphic,
which makes the resulting graphic harder to use in presentations to
other developers and to non-technical management.

This commit makes that replacement.