examples speedscope JSON output: correct units
This commit corrects the "unit" field in the austin.json example data
provided in the austin-python repository. Instances of the string
"\u03bcs" were replaced with the string "microseconds" so that the
resulting JSON file conforms to the Speedscope JSON schema, as
discussed in the commit message of the predecessor commit making a
similar replacement in the Speedscope Units enumeration.