Speedscope.dump method: correct docstring
Prior to this commit, the docstring for the Speedscope.dump method
read:

"Dump the pprof protobuf message to the given binary stream."

This message is clearly inapplicable because it refers to a different
output format (pprof) and a different type of stream input (the type
hint for the `stream` argument is TextIO, indicating that a binary
stream is *not* the intended stream type).

This commit corrects the docstring to refer to the proper output
format (JSON) and the proper stream input (text, not binary).