test/__init__.py: convert DATA_FILE to abspath
This commit converts the path stored as a string in the `DATA_FILE`
variable of `test/__init__.py` from a relative path to an absolute
path so that tests using this path can be run without the current
working directory being the root of the project. This behavior is less
brittle, and is thus useful for executing tests while changing the
current working directory to explore different parts of the project.